syncdeletes: delete orphaned mythtv recordings with no matching symlink
The purpose of this script is to allow user's of mythicalLibrarian
to delete the mythtv recordings that the mythicalLibrarian video
symlinks link to. Currently, if a video symlink is deleted it will no
longer show up in the user's "video library" but the original (and large)
video and all it's related files still exist in the mythtv database.

Finds all mythtv recordings that are no longer "attached" to a video
symlink (most likely generated by mythicalLibrarian). Recordings that
are still recording or have an active job are ignored. Any mythtv
recording that is "orphaned" is deleted.

Signed-off-by: Joe Swantek <joseph.swantek@gmail.com>