preshutdown: initial contribution
Used in the preshutdown sequence of a mythtv backend to test if it
is 'ok' to shutdown/suspend the machine. Checks if any user's are logged
into the machine, if a transcode is running (i.e. mythbrake) or if
mythicalLibrarian is running. If any of those items are active then
shutdown is is halted.

This can be simplified later by converting the shell script to python and
interfacing with the mythtv backend and checking its Job queue.

Signed-off-by: Joe Swantek <joseph.swantek@gmail.com>