myth_comm_cut: WIP - BROKEN
This is the start of a script that would snip out the non-commercial
bits or a recordings and stitch them back together...  The snip/stitch
part was functional in the past, but the chunks were almost never right.

I am not yet sure whether the bug(s) lie in commercial detection or the
snipping logic...

To debug that, I had been doing these tests:

Find the commercials:
mythcommflag -f /var/lib/mythtv/recordings/3197_20150915175000.mpg --outputmethod essentials --method=d2_all --outputfile commflag.edl --skipdb

Snip out one chunk:
/usr/bin/HandBrakeCLI -v -q 20.0 -e x264 -r 25 --crop 0:0:0:0 -d -m -x b-adapt=2:rc-lookahead=50:ref=3:bframes=3:me=umh:subme=8:trellis=1:merange=20:direct=auto -i "/var/lib/mythtv/recordings/3197_20150915175000.mpg" --start-at frame:6628 --stop-at frame:600 -o ~/snip.mp4 -4 --optimize 2

NOTE that the --stop-at is not what it sounds like... This appears to
be more of "do this many frames" than "where to stop"... Would have
been better called "--stop-after".  I am also not 100% clear on the
handbrake commands, but it seems to be correct.