mythbrake: drop DB connection before transcode and re-connect on it's completion
It is possible during a long transcode that the database connection
could timeout. This resulted in a MySQL failure ('MySQL server has gone
away') on the first access attempt of the MythTV database after the
transcode was complete. Now drop the database connection before the
transcode starts and re-establish it after the transcode completes to
ensure there are no timeout problems.

Signed-off-by: Joe Swantek <joseph.swantek@gmail.com>