mythbrake: delete the db reference before transcode instead of set to None
Setting to None relies on the python garbage collector to clear the
referenced object which isn't guaranteed to run and therefore still
result in connection issues. By deleting the object we are forcing
the MythDB reference to be removed and therefore the MythDB.shared
value to be cleared which is the flag triggering the use of a cached
connection or not. By 'shared' being cleared a re-connection to the DB
will create a new connection.

Signed-off-by: Joe Swantek <joseph.swantek@gmail.com>